[WIP] Implement custom string literals
TODO
* Test

Bugs:
It does not provide ways to implement %W{}. The behavior of %W{} is
confusing and not recommended for new literals. (Who thinks "#{'foo bar'}"
has one word 'foo bar', not two words, 'foo' and 'bar'?)
It does not allow to modify other builtins for consistency.

Made for a presentation in Fukuoka.rb 02 Zenya-sai.