Test link result by executing the output
The implementation since r50215 tested link results by checking if stderr
is empty or not. However, stderr may contain non-fatal warnings.
The warning of sprintf() by OpenBSD is an example of a non-fatal warning.

To distinguish such warnings from warnings fatal on the runtime
environment, this change lets test the link result by executing the output.

This change does not affect cross compilation.